+++ src/rank-font-data.js
@@ -1,0 +1,419 @@
+// Authentic Dota 2 RadianceM font glyph masks for rank plaques
+module.exports = {
+  "scale1": {
+    "0": {
+      "w": 19,
+      "h": 25,
+      "data": "AAAAAAAATpz///+uXAwAAAAAAAAAAAAmw///////////MwAAAAAAAAAm//////////////9AAAAAAAAAw////////////////xkAAAAAa///////MwAZw/////+LAAAAAP//////JgAAAAza/////xkAAED/////iwAAAAAATv////96AACL/////0AAAAAAAAz/////wwAAw/////8MAAAAAAAAw/////8AAP//////AAAAAAAAAK7/////Mxn/////rgAAAAAAAABr/////0Az/////64AAAAAAAAAa/////9rM/////+uAAAAAAAAAGv/////azP/////rgAAAAAAAABr/////2sm/////64AAAAAAAAAa/////9OAP//////AAAAAAAAAJz/////MwDD/////wAAAAAAAACu/////wwAi/////8zAAAAAAAA/////9oAAE7/////egAAAAAAQP////96AAAM//////8MAAAAAK7/////JgAAAGv/////wwwAAIv/////rgAAAAAM2v//////rv///////yYAAAAAACb//////////////04AAAAAAAAAJsP//////////0AAAAAAAAAAAAAATpza//+uXAwAAAAAAA=="
+    },
+    "1": {
+      "w": 11,
+      "h": 25,
+      "data": "AAAAAAAAGYv//2sAAAAAJov/////awAATq7///////9rXNr//////////2uc////////////a2v/////rv////9rM//aXAwz/////2sZTgAAADP/////awAAAAAAM/////9rAAAAAAAz/////2sAAAAAADP/////awAAAAAAM/////9rAAAAAAAz/////2sAAAAAADP/////awAAAAAAM/////9rAAAAAAAz/////2sAAAAAADP/////awAAAAAAM/////9rAAAAAAAz/////2sAAAAAADP/////awAAAAAAM/////9rAAAAAAAz/////2sAAAAAADP/////awAAAAAAM/////9rAAAAAAAz/////2s="
+    },
+    "2": {
+      "w": 17,
+      "h": 25,
+      "data": "AAAZXJzD////rnomAAAAAAAAev////////////+cDAAAAACu///////////////DDAAAAK7///////////////+LAAAArv//rmtri9r///////8ZAACuawwAAAAAAIv//////1wAACYAAAAAAAAAAK7/////nAAAAAAAAAAAAAAATv////+uAAAAAAAAAAAAAAAz/////64AAAAAAAAAAAAAADP/////awAAAAAAAAAAAAAAev////9AAAAAAAAAAAAAAAz//////wAAAAAAAAAAAAAAa/////9rAAAAAAAAAAAAACb/////2gwAAAAAAAAAAAAM2v////8zAAAAAAAAAAAAAIv/////awAAAAAAAAAAAABO/////4sAAAAAAAAAAAAAQP/////DDAAAAAAAAAAAACb/////wwwAAAAAAAAAAAAm//////8mAAAAAAAAAAAAJv//////XDMzMzMzMzMzACb//////////////////2sA////////////////////awD///////////////////9rJv///////////////////2s="
+    },
+    "3": {
+      "w": 15,
+      "h": 25,
+      "data": "AAxOi67///+uehkAAAAAAP////////////+cDAAAAP//////////////wwAAAP///////////////2sAAP/DazMzQIv//////9oAADMAAAAAAABO//////8ZAAAAAAAAAAAAw/////8zAAAAAAAAAAAArv////8ZAAAAAAAAAAAm/////8MAAAAAAAAAADP//////zMAAAAAJq6u2v//////TgAAAAAAM////////5wZAAAAAAAAM/////////96DAAAAAAAM///////////2iYAAAAAAAAADEDD/////9oMAAAAAAAAAAAM2v////9cAAAAAAAAAAAAev////+uAAAAAAAAAAAAa///////AAAAAAAAAAAAw///////DAwAAAAAAABO//////+uM/+LTjMzQIv///////9cM////////////////9oMM////////////////yYAM/////////////+uGQAAAEB6rv////+uejMAAAAA"
+    },
+    "4": {
+      "w": 19,
+      "h": 25,
+      "data": "AAAAAAAAAAAAAAAAQP+uAAAAAAAAAAAAAAAAAAAAJv//rgAAAAAAAAAAAAAAAAAADMP//64AAAAAAAAAAAAAAAAAAJz///+uAAAAAAAAAAAAAAAAAFz/////rgAAAAAAAAAAAAAAADP//////64AAAAAAAAAAAAAABn///////+uAAAAAAAAAAAAAADD////////rgAAAAAAAAAAAACL//+urv///64AAAAAAAAAAABO////DK7///+uAAAAAAAAAAAm////QACu////rgAAAAAAAAAM2v//egAArv///64AAAAAAAAArv//wwAAAK7///+uAAAAAAAAev///xkAAACu////rgAAAAAATv///0AAAAAArv///64AAAAAJv//////////////////////rq7//////////////////////66u//////////////////////+uTmtra2tra2tra9r////aa2trTgAAAAAAAAAAAACu////rgAAAAAAAAAAAAAAAAAArv///64AAAAAAAAAAAAAAAAAAK7///+uAAAAAAAAAAAAAAAAAACu////rgAAAAAAAAAAAAAAAAAArv///64AAAAAAAAAAAAAAAAAAK7////DAAAAAA=="
+    },
+    "5": {
+      "w": 16,
+      "h": 25,
+      "data": "AAAZMzMzMzMzMzMzM3ozAAAArv//////////////MwAAAK7//////////////zMAAACu//////////////8zAAAA2v//////////////MwAAAP////8mAAAAAAAAAAAAAAD/////AAAAAAAAAAAAAAAm/////wAAAAAAAAAAAAAAM////8MAAAAAAAAAAAAAAED///+uAAAAAAAAAAAAAABr/////////65cDAAAAAAAa/////////////9cAAAAAJz//////////////4sAAACuw66urtr/////////QAAAAAAAAAAAGXr//////8MAAAAAAAAAAAAAa///////GQAAAAAAAAAAAAz//////zMAAAAAAAAAAAAA//////8zAAAAAAAAAAAAGf//////MyYAAAAAAAAADMP//////wBr2npAMzMzev///////2sAa////////////////8MAAGv//////////////8MMAABr/////////////3oMAAAAAEB6rv/////DaxkAAAAAAA=="
+    },
+    "6": {
+      "w": 18,
+      "h": 25,
+      "data": "AAAAAAAAAAAm////////XAAAAAAAAAAAAE7//////8MZAAAAAAAAAAAATv//////iwwAAAAAAAAAAABO//////+LAAAAAAAAAAAAAED//////3oAAAAAAAAAAAAAJv//////egAAAAAAAAAAAAAArv////+LAAAAAAAAAAAAAABO/////64AAAAAAAAAAAAAAAz//////wwAAAAAAAAAAAAAAE7/////XE6LrpxrJgAAAAAAAMP//////////////5wMAAAAGf/////////////////DGQAAXP//////////////////wwAAi///////axkAGYv//////04Arv////9AAAAAAABr/////8MA/////4sAAAAAAAAA2v////8A/////1wAAAAAAAAAev////8z/////2sAAAAAAAAAa/////8zw////5wAAAAAAAAAnP////8Zev////8mAAAAAAAm/////9oAGf//////MwAAADP//////1wAAGv//////9qu2v//////wwAAAACL///////////////DDAAAAAAATv///////////2sAAAAAAAAAAAxOnNr/2pxcDAAAAAAA"
+    },
+    "7": {
+      "w": 18,
+      "h": 25,
+      "data": "/8OcTjMzMzMzMzMzMzMzQGsMw////////////////////1wArv///////////////////wwArv//////////////////nAAAa///////////////////TgAAAAAAAAAAAAAAAMP////aAAAAAAAAAAAAAAAAM/////9cAAAAAAAAAAAAAAAAnP////8MAAAAAAAAAAAAAAAZ/////3oAAAAAAAAAAAAAAACL/////xkAAAAAAAAAAAAAAAz/////iwAAAAAAAAAAAAAAAGv/////JgAAAAAAAAAAAAAADP////+uAAAAAAAAAAAAAAAATv////9AAAAAAAAAAAAAAAAA2v///9oAAAAAAAAAAAAAAABO/////04AAAAAAAAAAAAAAADa/////wwAAAAAAAAAAAAAADP/////awAAAAAAAAAAAAAAAK7/////DAAAAAAAAAAAAAAAM/////+LAAAAAAAAAAAAAAAArv////8ZAAAAAAAAAAAAAAAz/////64AAAAAAAAAAAAAAACu/////zMAAAAAAAAAAAAAADP/////2gAAAAAAAAAAAAAAAK7/////TgAAAAAAAAAAAAAA"
+    },
+    "8": {
+      "w": 17,
+      "h": 25,
+      "data": "AAAAAAAZXJyurnpADAAAAAAAAAAZnP//////////XAAAAAAAJv//////////////iwAAAAza////////////////XAAATv/////DMwAMa///////AACu/////wwAAAAAa/////8zAK7///+cAAAAAAAz/////zMArv///5wAAAAAACb/////MwB6/////xkAAAAAa////9oAABn/////2jMAAFz/////QAAAAE7//////9ra/////3oAAAAAACbD//////////9AAAAAAAAAa////////////8MmAAAADMP/////w67///////9AAACu/////04AAAxr//////8ZQP////9cAAAAAABO/////4uc/////wAAAAAAAADD/////67/////AAAAAAAAAIv/////rv////8MAAAAAAAAw/////+c/////4sAAAAAAE7/////2k7//////65AMzOL//////9rAMP/////////////////2gwADMP///////////////8mAAAADJz///////////+uGQAAAAAAACZ6w////8N6MwAAAAA="
+    },
+    "9": {
+      "w": 18,
+      "h": 25,
+      "data": "AAAAAAAMXJza/9qcTgwAAAAAAAAAAGv///////////9OAAAAAAAMw///////////////iwAAAACu///////artr//////1wAAE7//////zMAAABA//////8ZANr/////JgAAAAAAM/////96Gf/////DAAAAAAAAAMP////DM/////9rAAAAAAAAAGv/////M/////9rAAAAAAAAAGv/////DP/////DAAAAAAAAAIv/////AMP/////QAAAAAAAJv/////DAGv//////1wAAAAz//////+cAAz////////artr///////9cAAAz//////////////////8mAAAAJtr//////////////9oAAAAAAAxcrv//2ot6/////2sAAAAAAAAAAAAAAAzD/////wwAAAAAAAAAAAAAAIv/////awAAAAAAAAAAAAAAXP/////aDAAAAAAAAAAAAABO//////8mAAAAAAAAAAAAAE7//////04AAAAAAAAAAAAAXP//////egAAAAAAAAAAAACL//////9rAAAAAAAAAAAADMP//////04AAAAAAAAAAABO////////JgAAAAAAAAAA"
+    }
+  },
+  "scale2": {
+    "0": {
+      "w": 17,
+      "h": 22,
+      "data": "AAAAAABOnP//2oszAAAAAAAAAAAMw/////////96AAAAAAAADMP///////////+LAAAAAACL/////4trrv////9OAAAAGf////9cAAAAnP///9oAAAB6////2gAAAAAZ/////0AAANr///9rAAAAAADD////iwAZ/////zMAAAAAAGv////aADP/////GQAAAAAATv////8Aa/////8AAAAAAAAz/////zNr/////wAAAAAAADP/////M2v/////AAAAAAAAM/////8za/////8AAAAAAAAz/////zMz/////xkAAAAAAE7/////ABn/////MwAAAAAAa////9oAAP////9rAAAAAACu////iwAAi////8MAAAAAGf////9AAAAz/////1wAAACc////2gwAAACL/////4trrv////9OAAAAAAza////////////iwAAAAAAACbD/////////4sMAAAAAAAAAABOrv///5wzAAAAAAA="
+    },
+    "1": {
+      "w": 10,
+      "h": 22,
+      "data": "AAAAAAAATtr/rgAAAABO2v///64AGWv///////+uJv//////////rgz//////////64A2v+uTjP///+uAFwmAAAz////rgAAAAAAM////64AAAAAADP///+uAAAAAAAz////rgAAAAAAM////64AAAAAADP///+uAAAAAAAz////rgAAAAAAM////64AAAAAADP///+uAAAAAAAz////rgAAAAAAM////64AAAAAADP///+uAAAAAAAz////rgAAAAAAM////64AAAAAADP///+uAAAAAABc////rg=="
+    },
+    "2": {
+      "w": 15,
+      "h": 22,
+      "data": "AAxcnMP///+uXAwAAAAAJv////////////9AAAAAM///////////////TgAAM////////////////wwAM/+LMwAAGWv//////04AJiYAAAAAAABc/////4sAAAAAAAAAAAAA2v///64AAAAAAAAAAAAArv///64AAAAAAAAAAAAAw////2sAAAAAAAAAAAAZ/////0AAAAAAAAAAAACL////2gAAAAAAAAAAADP/////TgAAAAAAAAAADNr///+uAAAAAAAAAAAAnP////8ZAAAAAAAAAABr/////0AAAAAAAAAAAE7/////TgAAAAAAAAAAM/////9rAAAAAAAAAAAm/////4sAAAAAAAAAAE7//////66urq6urq5rQP////////////////9ra/////////////////9ra/////////////////9r"
+    },
+    "3": {
+      "w": 13,
+      "h": 22,
+      "data": "AECLrv///65cGQAAAGv///////////9OAABr/////////////0AAa///w67D////////AFxODAAAABnD/////zMAAAAAAAAAM/////8zAAAAAAAAAAz/////MwAAAAAAAABc/////wwAAAAAABlr/////04AAAAA/////////04AAAAAAP///////yYAAAAAAAD/////////egwAAAAAa2trw/////+uAAAAAAAAAABc/////04AAAAAAAAAAP////+cAAAAAAAAAADa////2gAAAAAAAAAz/////65cJgAAAAAmw/////+Lrv//rq7D////////M67/////////////egCu////////////awAAJmuu/////65cGQAAAA=="
+    },
+    "4": {
+      "w": 17,
+      "h": 22,
+      "data": "AAAAAAAAAAAAAAz//wAAAAAAAAAAAAAAAAAMw///AAAAAAAAAAAAAAAAAIv///8AAAAAAAAAAAAAAABc/////wAAAAAAAAAAAAAAQP//////AAAAAAAAAAAAACb///////8AAAAAAAAAAAAMw////////wAAAAAAAAAAAK7//2v/////AAAAAAAAAAB6//9rM/////8AAAAAAAAATv//nAAz/////wAAAAAAACb//9oMADP/////AAAAAAAM////JgAAM/////8AAAAADMP//1wAAAAz/////wAAAACL////////////////////a/////////////////////9r/////////////////////2sAAAAAAAAAADP/////AAAAAAAAAAAAAAAAM/////8AAAAAAAAAAAAAAAAz/////wAAAAAAAAAAAAAAADP/////AAAAAAAAAAAAAAAAM/////8AAAAAAAAAAAAAAAAz/////wAAAAA="
+    },
+    "5": {
+      "w": 14,
+      "h": 22,
+      "data": "ADP/////////////awAAM/////////////9rAAAz/////////////2sAAFz////Drq6urq6uTgAAa////zMAAAAAAAAAAAB6////AAAAAAAAAAAAAK7///8AAAAAAAAAAAAArv//2gAAAAAAAAAAAADa////////nE4AAAAAAP///////////9omAAAM//////////////8mABlcMzMzTov//////64AAAAAAAAAADP//////xkAAAAAAAAAAHr/////TgAAAAAAAAAAM/////9rAAAAAAAAAABA/////1wAAAAAAAAAAK7/////M3ozAAAAABmc/////9oArv//rq7D////////QACu/////////////4sAAK7///////////9cAAAAJmuu2v///65cDAAAAAA="
+    },
+    "6": {
+      "w": 17,
+      "h": 22,
+      "data": "AAAAAAAAAABO//////9cAAAAAAAAAAAAi//////DJgAAAAAAAAAAAIv/////nAwAAAAAAAAAAACL/////4sAAAAAAAAAAAAAev////+LAAAAAAAAAAAAADP/////iwAAAAAAAAAAAAAM2v///5wAAAAAAAAAAAAAAGv////aDAAAAAAAAAAAAAAA2v///4uL2v/aizMAAAAAAED//////////////3oAAAAAi////////////////4sAAADa//////+La4v//////zMAAP/////DDAAAACb/////nAAz/////xkAAAAAAE7/////DDP////DAAAAAAAADP////8zM////8MAAAAAAAAA/////zMA/////xkAAAAAAED/////DACc////wwwAAAAMw////64AABn//////3prev//////JgAAAE7//////////////04AAAAAAED//////////9ozAAAAAAAAAAxcrv///5xOAAAAAAA="
+    },
+    "7": {
+      "w": 16,
+      "h": 22,
+      "data": "Jv//rq6urq6urq6urq6urgD//////////////////04A/////////////////9oAAP////////////////9rAAAAAAAAAAAAABn/////DAAAAAAAAAAAAACL////iwAAAAAAAAAAAAAM/////xkAAAAAAAAAAAAAa////64AAAAAAAAAAAAAANr///8zAAAAAAAAAAAAAE7////aAAAAAAAAAAAAAADa////TgAAAAAAAAAAAAAz/////wAAAAAAAAAAAAAArv///2sAAAAAAAAAAAAAM/////8MAAAAAAAAAAAAAIv///+LAAAAAAAAAAAAABn/////GQAAAAAAAAAAAACL////rgAAAAAAAAAAAAAZ/////zMAAAAAAAAAAAAAi////9oAAAAAAAAAAAAAGf////9OAAAAAAAAAAAAAIv/////AAAAAAAAAAAAACb/////awAAAAAAAAAAAA=="
+    },
+    "8": {
+      "w": 15,
+      "h": 22,
+      "data": "AAAADE6c///anE4MAAAAAAAm2v/////////aJgAAACb//////////////yYAANr/////i2uL/////64AJv////8mAAAAM/////8mTv///5wAAAAAAK7///8zQP///4sAAAAAAK7///8zGf////8ZAAAAGf///9oAAHr/////TgxA/////0AAAACL////////////TgAAAAAAnP////////+LDAAAAAzD////////////2iYAAMP////DGQxO2v////8MTv///9oMAAAADMP///+Lrv///04AAAAAAED/////2v///zMAAAAAADP/////2v///4sAAAAAAE7/////i/////9OAAAAM/////+cM///////2q7D//////8zAHr//////////////4sAAABc////////////awAAAAAADFyu////rlwZAAAA"
+    },
+    "9": {
+      "w": 16,
+      "h": 23,
+      "data": "AAAAAAxcrv///5xOAAAAAAAAAED//////////8MZAAAAAE7//////////////xkAADP/////2npri//////DAACu////wwwAAAAm/////04m/////xkAAAAAAGv///+cXP///8MAAAAAAAAZ/////2v///+uAAAAAAAAAP////9c////2gAAAAAAAAz/////Jv////8zAAAAAABO////2gDa/////yYAAAAm/////64Aa///////i2uL//////96AADD////////////////QAAADK7//////////////wwAAAAATpz//8N6rv///4sAAAAAAAAAAAAAJv////8ZAAAAAAAAAAAADMP///+LAAAAAAAAAAAAAMP/////DAAAAAAAAAAAAIv/////QAAAAAAAAAAADIv/////XAAAAAAAAAAADMP/////iwAAAAAAAAAAJtr/////TgAAAAAAAAAMa///////QAAAAAAAAAA="
+    }
+  },
+  "scale3": {
+    "0": {
+      "w": 14,
+      "h": 19,
+      "data": "AAAAAFyu//+uTgAAAAAAAAzD////////wwwAAAAAw///////////wwAAAFz///+LDAyL////TgAA2v//2gAAAADa///aACb///9cAAAAAGv///8mXP///zMAAAAAM////1yL////DAAAAAAM////eq7///8AAAAAAAD///+urv///wAAAAAAAP///66u////AAAAAAAA////rov///8MAAAAAAz///+La////zMAAAAAM////2sm////XAAAAABc////JgDa///aAAAAANr//9oAAGv///+LAACL////awAADMP//////////8MAAAAAGcP////////DGQAAAAAADGvD///DawwAAAA="
+    },
+    "1": {
+      "w": 8,
+      "h": 19,
+      "data": "AAAAABmL//8AADOL/////07D////////rv////////96//96w////05ODACu////AAAAAK7///8AAAAArv///wAAAACu////AAAAAK7///8AAAAArv///wAAAACu////AAAAAK7///8AAAAArv///wAAAACu////AAAAAK7///8AAAAArv///wAAAACu////AAAAAK7///8="
+    },
+    "2": {
+      "w": 12,
+      "h": 19,
+      "data": "DE6cw///rnoZAAAAa///////////TgAAa////////////0AAa/+ua2uc/////8MAThkAAAAAQP////8ZAAAAAAAAAJz///8zAAAAAAAAAGv///8zAAAAAAAAAHr///8AAAAAAAAAANr//8MAAAAAAAAAXP///04AAAAAAAAZ////wwAAAAAAAADD////JgAAAAAAAIv///9OAAAAAAAATv///4sAAAAAAABO////iwAAAAAAAE7/////MzMzMzMzTv//////////////////////////////w///////////////"
+    },
+    "3": {
+      "w": 11,
+      "h": 19,
+      "data": "M3qu///aizMAAAD//////////3oAAP///////////04A/3pAM07D////wwAMAAAAAAz/////AAAAAAAAAK7///8AAAAAAAAM2v//wwAAAAAAGcP///8zAAAArv//////QAAAAACu/////3oMAAAAAK7///////8mAAAAAAAMXP///9oAAAAAAAAAev///04AAAAAAABr////awAAAAAAANr///9rrk4zM0Cu/////yb///////////+uAP//////////rgwAXK7/////nEAAAAA="
+    },
+    "4": {
+      "w": 15,
+      "h": 19,
+      "data": "AAAAAAAAAAAAAJz/AAAAAAAAAAAAAAAAa///AAAAAAAAAAAAAABA////AAAAAAAAAAAAACb/////AAAAAAAAAAAADMP/////AAAAAAAAAAAAnP//////AAAAAAAAAABr//+L////AAAAAAAAAED//0Br////AAAAAAAAJv//egBr////AAAAAAAMw//DAABr////AAAAAACL//8ZAABr////AAAAAFz//9qurq7a////rq6uDP//////////////////M///////////////////AAAAAAAAAABr////AAAAAAAAAAAAAABr////AAAAAAAAAAAAAABr////AAAAAAAAAAAAAABr////AAAAAAAAAAAAAABr////AAAA"
+    },
+    "5": {
+      "w": 12,
+      "h": 19,
+      "data": "AAAmMzMzMzMzM2szAACu//////////8zAADa//////////8zAAD///////////8zAAD///8AAAAAAAAAACb///8AAAAAAAAAADP//8MAAAAAAAAAAED//////65rDAAAAGv//////////zMAAGv///////////8ZACYAAAAZa/////+LAAAAAAAAAED/////AAAAAAAAAAD/////AAAAAAAAAAD/////DAAAAAAAAFz////DM8NcMzMzi/////9OM////////////8MAM///////////nAwADFyu/////5wzAAAA"
+    },
+    "6": {
+      "w": 14,
+      "h": 19,
+      "data": "AAAAAAAADMP///+uGQAAAAAAACba////egAAAAAAAAAm/////04AAAAAAAAADNr///9OAAAAAAAAAACu////TgAAAAAAAAAATv///4sAAAAAAAAAAADa///DAAAAAAAAAAAATv///4uu//+cMwAAAACu////////////iwAADP//////////////XAAz/////1wAAFz/////AFz///9cAAAAAFz///8za////wwAAAAADP///2tr////DAAAAAAA////ayb///9cAAAAAE7///9AAMP///9OAAAz////2gAAJv////////////8zAAAAM///////////QAAAAAAADGvD///DawwAAAA="
+    },
+    "7": {
+      "w": 14,
+      "h": 19,
+      "data": "a9qLMzMzMzMzMzMzTiZO///////////////aADP//////////////1wAGcP/////////////GQAAAAAAAAAAANr//64AAAAAAAAAAAAz////MwAAAAAAAAAAAK7//8MAAAAAAAAAAAAm////TgAAAAAAAAAAAIv//9oAAAAAAAAAAAAZ////awAAAAAAAAAAAIv///8MAAAAAAAAAAAM////egAAAAAAAAAAAGv///8ZAAAAAAAAAAAM////iwAAAAAAAAAAAGv///8mAAAAAAAAAAAM////rgAAAAAAAAAAAGv///9AAAAAAAAAAAAM////2gAAAAAAAAAAAGv///9OAAAAAAAAAAA="
+    },
+    "8": {
+      "w": 14,
+      "h": 19,
+      "data": "AAAADFyu//+uawwAAAAAACba/////////yYAAAAM2v///////////wwAAE7///+LDAB6////XAAAi///2gAAAADD//+uAACc//+uAAAAAK7//5wAAE7///8mAAAZ////TgAAAMP///9rTtr//5wAAAAADK7///////+LDAAAAAAZrv///////8MmAAAAGf///9pci/////8mAACc///DDAAAGdr//64ADP///04AAAAATv///xkz////MwAAAAAz////Mxn///+LAAAAAGv///8mANr///9rMzNr////2gAAQP////////////9AAAAATv//////////TgAAAAAAGYva///aixkAAAA="
+    },
+    "9": {
+      "w": 14,
+      "h": 19,
+      "data": "AAAAGXra///DawwAAAAAAE7//////////zMAAABO/////////////yYAGf///8MmAABc////wwBr////GQAAAABc////Jq7//64AAAAAAAz///9rrv//wwAAAAAAAP///2tr////MwAAAABc////XCb/////QAAAXP////8zAJz//////////////wwADMP///////////+cAAAAAE6u//+cnP///04AAAAAAAAAAAza///aAAAAAAAAAAAAi////0AAAAAAAAAAAIv///+uAAAAAAAAAABr////wwwAAAAAAAAAi/////8mAAAAAAAADIv////DGQAAAAAAADPD////wwwAAAAAAAA="
+    }
+  },
+  "scale4": {
+    "0": {
+      "w": 11,
+      "h": 16,
+      "data": "AAAAABkzGQAAAAAAACbD////wyYAAAAm/////////yYAANr//1wAXP//2gBA///DAAAAw///QIv//2sAAABr//+Lw///MwAAADP//8P///8zAAAAM////////zMAAAAz////////MwAAADP///+u//9AAAAAQP//rnr//3oAAAB6//96Jv//2gAAANr//yYArv//nDOc//+uAAAMw///////wwwAAAAMa8P/w2sMAAA="
+    },
+    "1": {
+      "w": 7,
+      "h": 15,
+      "data": "AAAADGva/wAmi/////9O////////M//aa////wxOAAD///8AAAAA////AAAAAP///wAAAAD///8AAAAA////AAAAAP///wAAAAD///8AAAAA////AAAAAP///wAAAAD///8AAAAA////"
+    },
+    "2": {
+      "w": 11,
+      "h": 16,
+      "data": "ABlcnK6uaxkAAAAA/////////1wAAAD//////////0AAAK5AAABO////rgAMAAAAAABr////AAAAAAAAADP///8AAAAAAAAAQP//rgAAAAAAAACL//9cAAAAAAAAM////wwAAAAAAAzD//9cAAAAAAAAi///nAAAAAAAAE7//8MMAAAAAABO////DAAAAAAATv///5xra2trMyb///////////9rM////////////2s="
+    },
+    "3": {
+      "w": 10,
+      "h": 16,
+      "data": "ACZOa2szAAAAAJz///////9AAACu/////////xkAa0AAAED///9rAAAAAAAAev//rgAAAAAAAIv//2sAAAAAAFz///8ZAAAA/////64mAAAAAP//////awAAAAAzM07a//9rAAAAAAAAGf///wAAAAAAAAz///8zDAAAAABr////DGvDemuu////rgBr/////////yYAM5zD//+uawwAAA=="
+    },
+    "4": {
+      "w": 13,
+      "h": 15,
+      "data": "AAAAAAAAAACcrgAAAAAAAAAAAABr/64AAAAAAAAAAABO//+uAAAAAAAAAAAm////rgAAAAAAAAAM/////64AAAAAAAAMw/9A//+uAAAAAAAAi/9rAP//rgAAAAAAa/+cAAD//64AAAAATv//QDMz///DMzMMGf//////////////MzP//////////////zMAAAAAAAAA//+uAAAAAAAAAAAAAP//rgAAAAAAAAAAAAD//64AAAAAAAAAAAAA//+uAAAA"
+    },
+    "5": {
+      "w": 10,
+      "h": 15,
+      "data": "AHr///////+uAACu////////rgAArv/aa2tra04AAMP/rgAAAAAAAAD//3oAAAAAAAAA//+LMyYAAAAAJv//////2lwAADP/////////awAMDAAAJq7///8MAAAAAAAM////MwAAAAAAAMP//04MAAAAAED///8mrsNra5z///+uAK7////////DDABAnNr//65OAAAA"
+    },
+    "6": {
+      "w": 11,
+      "h": 15,
+      "data": "AAAAAACL////TgAAAAAMw///wwwAAAAADMP//4sAAAAAAACu//+LAAAAAAAATv//iwAAAAAAAADa//+c///DTgAAJv//////////iwBr////azNr////Tpz//04AAABc//+urv//AAAAAAD///+u//8MAAAAAP///1z//1wAAABO//+cDP///3oza////yYAJv////////9OAAAAGXqu/656GQAA"
+    },
+    "7": {
+      "w": 11,
+      "h": 15,
+      "data": "rv///////////5yu////////////GUBra2tra5z//4sAAAAAAAAArv//JgAAAAAAABn//64AAAAAAAAAi///QAAAAAAAABn//9oAAAAAAAAAa///TgAAAAAAAAz///8MAAAAAAAAa///awAAAAAAAADa//8ZAAAAAAAATv//iwAAAAAAAADa//8ZAAAAAAAATv//rgAAAAAAAAza//9AAAAAAAAA"
+    },
+    "8": {
+      "w": 11,
+      "h": 16,
+      "data": "AAAAABkzJgAAAAAAADPD/////1wAAAAz/////////2sAAMP//04AM////wwA//+uAAAAev//MwD//9oAAACL//8MAGv//5wzXP//egAAAGv//////4sAAAAMnP///////04AAMP//1wATv///0BA//+LAAAAM///rmv//2sAAAAA////XP//wwAAAFz//8MM////2muL////TgBO/////////4sAAAAZesP//5wzAAA="
+    },
+    "9": {
+      "w": 11,
+      "h": 16,
+      "data": "AAAZeq7/rmsMAAAATv////////8mACb///9rM4v///8MnP//XAAAAGv//1z///8AAAAADP//rv///wAAAAAA//+uw///QAAAAED//5xr////MwAz////awzD//////////9AAAyL////2v///wwAAAAMMyZc//+LAAAAAAAAJv///xkAAAAAACb///9cAAAAAAAm////iwAAAAAAJv///4sAAAAADGv///9rAAAAAAA="
+    }
+  },
+  "mini1": {
+    "0": {
+      "w": 9,
+      "h": 12,
+      "data": "AAAMXK6cTgAAAAzD/////4sAAGv/wwwm//8zANr/QAAAev+LDP//DAAAQP/DM///AAAAM///M///AAAAM///Gf//AAAAM//aAP//MwAAa/+cAIv/nAAM2v9OABn//8Pa/8MAAAAmnP//iwwA"
+    },
+    "1": {
+      "w": 6,
+      "h": 11,
+      "data": "AAAMa9quGYv///+uAP+u2v+uACYArv+uAAAArv+uAAAArv+uAAAArv+uAAAArv+uAAAArv+uAAAArv+uAAAArv+u"
+    },
+    "2": {
+      "w": 8,
+      "h": 11,
+      "data": "M67//8NOAABr//////9rAEAmAAzD/9oAAAAAADP//wAAAAAATv/DAAAAAADD/2sAAAAAev/aDAAAAE7//yYAAABO//9OAAAATv///66urmuL////////aw=="
+    },
+    "3": {
+      "w": 8,
+      "h": 11,
+      "data": "GZz//8NOAAAz//////9rABkmAAyc/64AAAAAAJz/iwAAAK7a/64MAAAA////2iYAAAAADIv//wwAAAAAM///MxkZAACL//8mM///////rgAZrv//2msMAA=="
+    },
+    "4": {
+      "w": 9,
+      "h": 11,
+      "data": "AAAAAAAMwzMAAAAAAACc/zMAAAAAAHr//zMAAAAATv/a/zMAAAAm/06u/zMAABn/iwCu/zMADMP/a2va/4trQP//////////AAAAAACu/zMAAAAAAACu/zMAAAAAAACu/zMA"
+    },
+    "5": {
+      "w": 8,
+      "h": 11,
+      "data": "ADP//////wAAXP//////AABr/1wAAAAAAGv/MwAAAAAArv///5wZAACLrq7///8MAAAAACb//04AAAAAAK7/axkmAABc//9AM///////rgAZrv//2msMAA=="
+    },
+    "6": {
+      "w": 8,
+      "h": 11,
+      "data": "AAAATv//awwAAE7//04AAABA//8mAAAADP//QAAAAABc///D/8NAAK7//67a//9A//9OAACL/8P//wAAADP//8P/TgAAi/+cM///rsP//yYAJpz//4sZAA=="
+    },
+    "7": {
+      "w": 8,
+      "h": 11,
+      "data": "/////////2v/////////GQAAAABr/4sAAAAADP//GQAAAABr/64AAAAAANr/MwAAAABO/64AAAAAANr/TgAAAABO/9oAAAAAANr/TgAAAABc//8MAAAAAA=="
+    },
+    "8": {
+      "w": 8,
+      "h": 12,
+      "data": "AAxcrq5cAAAMw//////DAFz/wwwZ2v8za/9rAACL/0Az/9ozJv//DABc/////yYAJtr/rtr/wwyu/1wAAJz/i///MwAAM//D//+LAAyc/5xr////////MwBOw///nDMA"
+    },
+    "9": {
+      "w": 9,
+      "h": 12,
+      "data": "AABAnP//iwwAAE7//8Pa/8MMAP//XAAMw/9cM///AAAATv+uM///DAAAXP+uAP//awAMrv+LAGv//8Pa//9cAABr2v/D//8ZAAAAAACL/5wAAAAAAIv//xkAAAAAi///JgAAACbD/9omAAAA"
+    }
+  },
+  "mini2": {
+    "0": {
+      "w": 9,
+      "h": 11,
+      "data": "AAAMXK56GQAAAAzD/////yYAAIv/nAxc//8AAP//MwAA//8zJv//DAAArv9cM///DAAArv96M///DAAArv96AP//JgAA2v9AAJz/egAz//8MACb//8P//2sAAAAz2v//XAwA"
+    },
+    "1": {
+      "w": 6,
+      "h": 11,
+      "data": "AAAZi/9AGa7///9AJv+c//9AACYA//9AAAAA//9AAAAA//9AAAAA//9AAAAA//9AAAAA//9AAAAA//9AAAAA//9A"
+    },
+    "2": {
+      "w": 8,
+      "h": 11,
+      "data": "GWuurnoZAABr//////8ZAEBAACb//5wAAAAAAGv/wwAAAAAAev+LAAAAAAz//yYAAAAAi/+LAAAAAE7/wwwAAAAm/9oZAAAAQP///66urgCu////////DA=="
+    },
+    "3": {
+      "w": 7,
+      "h": 11,
+      "data": "DGucrnoZADP//////xkmQAAm//9rAAAAAK7/egAAADP//yYAAP///04AAACuw///GQAAAAB6/64MAAAAnP/DM/+u2v//Thmu//+uTgA="
+    },
+    "4": {
+      "w": 9,
+      "h": 11,
+      "data": "AAAAAABOwwAAAAAAACb/wwAAAAAADNr/wwAAAAAArtr/wwAAAAB62k7/wwAAAE7/MzP/wwAAJv/aa4v//2sza/////////96AAAAADP/wwAAAAAAADP/wwAAAAAAADP/wwAA"
+    },
+    "5": {
+      "w": 8,
+      "h": 11,
+      "data": "AE6urq6uawAArv////96AACu/wwAAAAAAMP/DAAAAAAA//+uiyYAAAD//////0AADAwAGcP/2gAAAAAAM///DBkAAACL/8MMa/+u2v//TgAzrv//nEAAAA=="
+    },
+    "6": {
+      "w": 9,
+      "h": 11,
+      "data": "AAAAAIv//0AAAAAAi//DGQAAAAB6/8MZAAAAADP//yYAAAAAAJz//9r/egwAAP///67//5wAJv//QAAm//8ZM//aDAAArv9ADP//MwAM//8mAHr//67a/4sMAABc2v/aawwA"
+    },
+    "7": {
+      "w": 8,
+      "h": 11,
+      "data": "/66urq6urhn///////+cAAAAAADa/zMAAAAAM/+uDAAAAACu/1wAAAAAGf/aDAAAAACL/3oAAAAADP//GQAAAABr/5wAAAAADP//JgAAAABr/64MAAAAAA=="
+    },
+    "8": {
+      "w": 8,
+      "h": 11,
+      "data": "ABl6roszAAAm//////9OAJz/awxc/8MArv9AABn/wwxA//9cw/9OAACc////wwwAev+LM5z/wwD//wwAANr/Gf//MwAm//8Mev//rv//iwwAa9r/2msMAA=="
+    },
+    "9": {
+      "w": 9,
+      "h": 11,
+      "data": "AABr2v//XAAAAHr//67//4sAGf//GQAm//8ZM//DAAAAw/9ADP//JgAm//8zAIv//67///8MAAx6///a/64AAAAAAAz//0AAAAAADMP/iwAAAAAMw/+cDAAAADPa/4sMAAAA"
+    }
+  },
+  "mini3": {
+    "0": {
+      "w": 8,
+      "h": 9,
+      "data": "AAyu//9OAAAArv+L2v9AACb/wwxA/64AXP96AAD//wBr/3oAAP//DFz/egAA//8MJv/aAED/rgAArv+Lw/9AAAAMrv//XAAA"
+    },
+    "1": {
+      "w": 5,
+      "h": 9,
+      "data": "ACaL/0B6////QE5cw/9AAACu/0AAAK7/QAAArv9AAACu/0AAAK7/QAAArv9A"
+    },
+    "2": {
+      "w": 6,
+      "h": 9,
+      "data": "GU5rTgwArv///9oMXBkMw/9rAAAAa/9rAAAAw/8mAABr/3oAAE7/iwwAQP//emtO///////D"
+    },
+    "3": {
+      "w": 6,
+      "h": 9,
+      "data": "a9r/2kAAnHp6//8mAAAAw/9AAAAm//8ZAK7//04AACZAw/8zAAAAev96Tnp6//9ATv//2msM"
+    },
+    "4": {
+      "w": 8,
+      "h": 9,
+      "data": "AAAAAGvDAAAAAAAz/8MAAAAADP//wwAAAACu2v/DAAAAXP9O/8MAACb/i1z/2jMAnP///////wwAAAAz/8MAAAAAADP/wwAA"
+    },
+    "5": {
+      "w": 6,
+      "h": 9,
+      "data": "AGtra4tAJv////9AM/96AAAAM/+LJgAAa////9oMAAAMnP9rAAAAa/96rnp6//9Aev//2k4A"
+    },
+    "6": {
+      "w": 8,
+      "h": 9,
+      "data": "AAAAev/DMwAAAFz/wxkAAAAZ/9oZAAAAAIv/nGsZAAAA//////9OACb/nBlO/9oAM/96AAz//wwAw/96rv+cAAAZnP//ehkA"
+    },
+    "7": {
+      "w": 8,
+      "h": 9,
+      "data": "M/9ra2tregwm//////9rAAAAAADa/xkAAAAATv+cAAAAAACu/yYAAAAAM/+uAAAAAACu/0AAAAAAM/+uDAAAAACu/1wAAAAA"
+    },
+    "8": {
+      "w": 7,
+      "h": 9,
+      "data": "AAyc//+LDACc/1yL/2sArv8MJv96AFz/rsP/JgBA////2gwZ/9oZXP+cM/+cABn/wxn//4vD/4sAM8P//3oZ"
+    },
+    "9": {
+      "w": 8,
+      "h": 9,
+      "data": "ABmc//96AAAM2v96rv+LAE7/iwAM2v8AXP+LAAz//wwZ//+Lrv/DDABA2v///4sAAAAADMP/JgAAAAzD/1wAAAAz2v9cAAAA"
+    }
+  },
+  "mini4": {
+    "0": {
+      "w": 7,
+      "h": 8,
+      "data": "AAAZMwAAAABO////GQAZ/5wz/5wATv9OAMP/AGv/QACu/wxA/04Aw8MMDP/DXP96AAAm2v96DAA="
+    },
+    "1": {
+      "w": 4,
+      "h": 7,
+      "data": "GXr/esPa/3omM/96ADP/egAz/3oAM/96ADP/eg=="
+    },
+    "2": {
+      "w": 6,
+      "h": 8,
+      "data": "AAwzJgAAAP///9oMAE4Mev9rAAAAQP9cAAAArv8mAABc/2sAAE7/w0AmDP/////D"
+    },
+    "3": {
+      "w": 5,
+      "h": 8,
+      "data": "ADMzAACu////DEAMXP9rAABr/0AA//+cAAAAXP9cQDOL/1yL//96DA=="
+    },
+    "4": {
+      "w": 7,
+      "h": 7,
+      "data": "AAAAM/8MAAAAJv//DAAAJv/a/wwADMNcrv8MAIv//////0AAAACu/wwAAAAArv8MAA=="
+    },
+    "5": {
+      "w": 5,
+      "h": 8,
+      "data": "GTMzMwBr////QIvDDAAArv//nAAmAHr/awAAM/96ejOu/07D//9rDA=="
+    },
+    "6": {
+      "w": 7,
+      "h": 7,
+      "data": "AABr/5wzAAB6/1wAAAAZ////rhkAXP+LTv+uAGv/GQCu/wwm/5xA/64MAE7a/4sZAA=="
+    },
+    "7": {
+      "w": 7,
+      "h": 8,
+      "data": "ABkzMzMmAED/////nAwMGQBr/yYAAAAM/5wAAAAAXP9AAAAAANquDAAAAE7/XAAAAAza2gwAAAA="
+    },
+    "8": {
+      "w": 6,
+      "h": 8,
+      "data": "AAAZMwAAAIv///8mDP96DP96ANrDXP9OAIv///8mM/9cDP+uJv+uQP+uAGv//5wZ"
+    },
+    "9": {
+      "w": 6,
+      "h": 8,
+      "data": "AE7a/3oAM/+cQP9ra/9AAK7DQP+LDNquAJz///9rAAAZrv8mAABO/3oADGv/iwwA"
+    }
+  }
+};